requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
langfuse>=2.0.0
faiss-cpu>=1.8.0
openai>=1.0.0
streamlit>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
sqlalchemy>=2.0.0
pandas>=2.0.0
